# openapi.yaml
...
openapi: 3.1.0
info:
  title: EcomGen API
  version: 0.1.0
  description: Backend-first contract for the EcomGen desktop-first local application.
security: []
servers:
  - url: http://localhost:8787/api/v1
paths:
...
  /models:
    get:
      operationId: listModels
      responses:
        "200":
          description: Global model cast library; each entry carries its selected master portrait when one was picked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
    post:
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelInput"
      responses:
        "201":
          description: Model spec saved; no generation is triggered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcomModel"
        "400":
          $ref: "#/components/responses/ValidationError"
  /models/{modelId}:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    get:
      operationId: getModel
      responses:
        "200":
          description: Single model with its spec and selected portrait.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcomModel"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelInput"
      responses:
        "200":
          description: Updated model; changing the spec keeps existing portraits but future casts compile from the new spec.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcomModel"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteModel
      responses:
        "204":
          description: Model deleted together with its portraits and reference face files.
  /models/{modelId}/reference-face:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    post:
      operationId: uploadModelReferenceFace
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadModelReferenceFaceInput"
      responses:
        "201":
          description: Reference face stored; replaces any previous one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcomModel"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteModelReferenceFace
      responses:
        "204":
          description: Reference face removed; future casts fall back to text-only identity.
  /models/{modelId}/cast-jobs:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    post:
      operationId: createModelCastJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelCastJobInput"
      responses:
        "202":
          description: Cast job accepted (or an existing job reused for an identical request).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/CapabilityUnsupported"
  /models/{modelId}/portraits:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    get:
      operationId: listModelPortraits
      responses:
        "200":
          description: All portraits generated for the model, newest last.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPortraitList"
        "404":
          $ref: "#/components/responses/NotFound"
  /model-portraits/{portraitId}:
    parameters:
      - $ref: "#/components/parameters/ModelPortraitId"
    delete:
      operationId: deleteModelPortrait
      responses:
        "204":
          description: Portrait file and row removed.
        "404":
          $ref: "#/components/responses/NotFound"
  /model-portraits/{portraitId}/select:
    parameters:
      - $ref: "#/components/parameters/ModelPortraitId"
    post:
      operationId: selectModelPortrait
      responses:
        "200":
          description: Portrait marked as the model's master portrait; any previous selection is cleared. Returns the whole model so callers refresh selection state in one round trip.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcomModel"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    ModelId:
      name: modelId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ModelPortraitId:
      name: portraitId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Request does not match the schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    CapabilityUnsupported:
      description: The selected model lacks a requested capability.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ImageResolution:
      type: string
      enum:
        - 1K
        - 2K
        - 4K
    ImageAspectRatio:
      type: string
      enum:
        - AUTO
        - 1:1
        - 2:3
        - 3:2
        - 3:4
        - 4:3
        - 4:5
        - 5:4
        - 9:16
        - 16:9
        - 21:9
    ImageQuality:
      type: string
      enum:
        - auto
        - low
        - medium
        - high
    ImageOutputFormat:
      type: string
      enum:
        - png
        - jpeg
        - webp
...
    JobType:
      type: string
      enum:
        - PLAN
        - COPYWRITE
        - GENERATE
        - EXPORT
        - EDIT_PLAN
        - EDIT_GENERATE
        - LAYER_PLAN
        - LAYER_EXPORT
        - SUITE_FORGE
        - MODEL_CAST
        - PATTERN_EXTRACT
        - PATTERN_FORGE
        - PATTERN_DERIVE
        - PATTERN_VARIANT
        - PATTERN_TILE_CHECK
        - PRINT_PACK
        - PATTERN_DRAFT_GENERATE
        - PATTERN_DRAFT_EDIT
        - PATTERN_DRAFT_CUTOUT
        - PATTERN_DRAFT_PROCESS
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                patternProperties:
                  ^(.*)$: {}
            requestId:
              format: uuid
              type: string
...
    Job:
      type: object
      required:
        - id
        - type
        - status
        - progress
        - retryable
        - createdAt
      properties:
        id:
          format: uuid
          type: string
        type:
          $ref: "#/components/schemas/JobType"
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - SUCCEEDED
            - FAILED
            - CANCELLED
        progress:
          minimum: 0
          maximum: 100
          type: integer
        retryable:
          type: boolean
        requestFingerprint:
          anyOf:
            - type: string
            - type: "null"
        providerId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        modelId:
          anyOf:
            - type: string
            - type: "null"
        estimatedCost:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        actualCost:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        cancelRequested:
          type: boolean
        error:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        progressDetail:
          anyOf:
            - $ref: "#/components/schemas/JobProgressDetail"
            - type: "null"
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
    JobProgressDetail:
      type: object
      required:
        - shotsGenerated
        - shotsTarget
      properties:
        shotsGenerated:
          minimum: 0
          description: 流式观察到的已生成分镜数（模型可能回填或重试，不保证单调）。
          type: integer
        shotsTarget:
          description: 建单时指定的目标分镜数；未指定时为 null，此时只有分子没有分母。
          anyOf:
            - minimum: 1
              type: integer
            - type: "null"
...
    UploadModelReferenceFaceInput:
      type: object
      required:
        - file
      properties:
        file:
          format: binary
          description: 同一人正面清晰人像，仅支持图片文件。
          type: string
    ModelSpec:
      description: 四层结构化模特规格：身份内核 / 容貌细节 / 表达层 / 镜头呈现。定妆照 prompt 由该规格确定性编译；维度之间存在互斥约束，判定单源在 packages/ecom-skill。
      type: object
      required:
        - gender
        - age
        - heritage
        - stature
        - build
        - faceShape
        - eyeShape
        - eyeColor
        - browShape
        - noseShape
        - lipShape
        - hairLength
        - hairstyle
        - hairColor
        - hairTexture
        - hairline
        - complexion
        - skinTexture
        - facialHair
        - distinctiveMarks
        - expression
        - gaze
        - aura
        - makeup
        - baseWardrobe
        - framing
        - pose
        - backdrop
        - lighting
        - lens
      properties:
        gender:
          type: string
          enum:
            - FEMALE
            - MALE
            - ANDROGYNOUS
        age:
          type: string
          enum:
            - CHILD_7
            - PRETEEN_11
            - TEEN_16
            - EARLY_20S
            - LATE_20S
            - EARLY_30S
            - MID_30S
            - MID_40S
            - SENIOR
        heritage:
          type: string
          enum:
            - EAST_ASIAN
            - SOUTHEAST_ASIAN
            - SOUTH_ASIAN
            - MIDDLE_EASTERN
            - NORTHERN_EUROPEAN
            - MEDITERRANEAN
            - LATIN_AMERICAN
            - AFRICAN
            - MIXED
        stature:
          type: string
          enum:
            - PETITE_158
            - STANDARD_165
            - TALL_172
            - RUNWAY_180
        build:
          type: string
          enum:
            - SLENDER
            - BALANCED
            - ATHLETIC
            - MUSCULAR
            - CURVY
            - PLUS
            - MATERNITY
        faceShape:
          type: string
          enum:
            - OVAL
            - ROUND
            - SQUARE_JAW
            - HEART
            - LONG
            - DIAMOND
        eyeShape:
          type: string
          enum:
            - ALMOND
            - ROUND
            - MONOLID
            - UPTURNED
            - DOWNTURNED
            - NARROW
            - DEEP_SET
            - PEACH_BLOSSOM
            - INNER_DOUBLE
        eyeColor:
          type: string
          enum:
            - DARK_BROWN
            - AMBER
            - HAZEL
            - GREEN
            - BLUE
            - GREY
        browShape:
          type: string
          enum:
            - STRAIGHT_SOFT
            - ARCED
            - FEATHERED
            - BOLD
            - THIN_ARCHED
            - WILLOW
            - THICK_STRAIGHT
        noseShape:
          type: string
          enum:
            - DELICATE
            - STRAIGHT
            - SCULPTED
            - BROAD
            - UPTURNED
            - AQUILINE
            - BULBOUS
        lipShape:
          type: string
          enum:
            - NATURAL
            - FULL
            - BOW
            - WIDE
            - THIN
        hairLength:
          type: string
          enum:
            - CROP
            - CHIN_BOB
            - SHOULDER
            - LONG
            - WAIST
        hairstyle:
          type: string
          enum:
            - SLEEK_STRAIGHT
            - SOFT_WAVE
            - DEEP_CURL
            - HIGH_BUN
            - PONYTAIL
            - TEXTURED_FRINGE
            - BRAIDED
            - AFRO
            - SHAVED
            - PIXIE
            - SLICKED_BACK
            - LAYERED_LONG
            - HIME_CUT
            - HALF_UP
            - TWIN_TAILS
            - SPACE_BUNS
            - LOW_BUN
        hairColor:
          type: string
          enum:
            - INK_BLACK
            - ESPRESSO
            - CHESTNUT
            - LIGHT_BROWN
            - HONEY_BLONDE
            - STRAWBERRY_BLONDE
            - ASH_BROWN
            - COPPER
            - BURGUNDY
            - PLATINUM
            - SILVER
            - MILK_TEA
            - ASH_BLONDE
        hairTexture:
          type: string
          enum:
            - SLEEK_GLOSSY
            - NATURAL_VOLUME
            - TOUSLED
            - WET_LOOK
        hairline:
          type: string
          enum:
            - ROUNDED
            - SQUARE
            - WIDOWS_PEAK
            - RECEDING
            - HIGH_FOREHEAD
        complexion:
          type: string
          enum:
            - PORCELAIN
            - FAIR_WARM
            - LIGHT_NEUTRAL
            - MEDIUM_OLIVE
            - HONEY
            - CARAMEL
            - DEEP
            - DEEP_EBONY
        skinTexture:
          type: string
          enum:
            - NATURAL_PORES
            - DEWY
            - MATTE_FINE
            - FRECKLED
            - MATURE_LINES
        facialHair:
          type: string
          enum:
            - NONE
            - STUBBLE
            - MUSTACHE
            - GOATEE
            - BEARD
            - FULL_BEARD
        distinctiveMarks:
          maxItems: 2
          type: array
          items:
            type: string
            enum:
              - FRECKLES_NOSE
              - FRECKLES_CHEEKS
              - MOLE_CHEEK
              - TEAR_MOLE
              - DIMPLES
              - BEAUTY_MARK_LIP
              - BIRTHMARK
              - SUBTLE_ASYMMETRY
              - AEGYO_SAL
              - BROW_SCAR
        expression:
          type: string
          enum:
            - CALM_DIRECT
            - SOFT_SMILE
            - BRIGHT_GRIN
            - WARM_LAUGH
            - EDITORIAL_DISTANCE
            - THOUGHTFUL
            - TENDER
            - SURPRISED
        gaze:
          type: string
          enum:
            - DIRECT_TO_CAMERA
            - OFF_CAMERA
            - LOWERED_LIDS
            - HALF_LID_IDLE
        aura:
          maxItems: 3
          type: array
          items:
            type: string
            enum:
              - WARM_APPROACHABLE
              - QUIET_CONFIDENCE
              - METROPOLITAN
              - PLAYFUL
              - SERENE
              - ARTSY
              - POLISHED
              - FRESH
        makeup:
          type: string
          enum:
            - NONE
            - MINIMAL_DEWY
            - EVERYDAY
            - SOFT_GLAM
            - RED_LIP
            - SMOKY
            - MATURE
        baseWardrobe:
          type: string
          enum:
            - WHITE_TANK
            - BLACK_TURTLENECK
            - WHITE_SHIRT
            - NEUTRAL_TEE
            - BLAZER_SHIRT
            - RIBBED_DRESS
            - SLEEVELESS_KNIT
            - HOODIE
        framing:
          type: string
          enum:
            - FULL_BODY
            - THREE_QUARTER
            - WAIST_UP
            - BEAUTY_CLOSEUP
        pose:
          type: string
          enum:
            - WEIGHT_LEFT_HIP
            - NATURAL_WALK
            - LEANING_WALL
            - SEATED_STOOL
            - HANDS_RELAXED
            - HANDS_ON_HIPS
            - ARMS_CROSSED
            - OVER_SHOULDER
            - HANDS_IN_POCKETS
            - HAIR_TOUCH
        backdrop:
          type: string
          enum:
            - SEAMLESS_GREY
            - PURE_WHITE
            - CREAM
            - WARM_BEIGE
            - OUTDOOR_BOKEH
            - PALE_PINK
            - SLATE_BLUE
            - CHARCOAL
        lighting:
          type: string
          enum:
            - SOFTBOX_THREE_POINT
            - WINDOW_DAYLIGHT
            - OVERCAST_DIFFUSED
            - GOLDEN_BACKLIT
            - RING_LIGHT
            - DRAMATIC_SIDE
            - HIGH_KEY_STUDIO
            - BUTTERFLY_LIGHT
        lens:
          type: string
          enum:
            - LENS_35
            - LENS_50
            - LENS_85
            - LENS_105
...
    ModelPortrait:
      type: object
      required:
        - id
        - modelId
        - providerId
        - imageModelId
        - aspectRatio
        - selected
        - createdAt
      properties:
        id:
          format: uuid
          type: string
        modelId:
          format: uuid
          type: string
        url:
          description: 定妆照静态访问地址；仅列表与详情响应携带。
          type: string
        width:
          anyOf:
            - type: integer
            - type: "null"
        height:
          anyOf:
            - type: integer
            - type: "null"
        providerId:
          format: uuid
          type: string
        imageModelId:
          type: string
        aspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        selected:
          description: 定妆照标记；每个模特至多一张被选定。
          type: boolean
        createdAt:
          format: date-time
          type: string
    EcomModel:
      type: object
      required:
        - id
        - name
        - spec
        - notes
        - hasReferenceFace
        - portraitCount
        - createdAt
        - updatedAt
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        spec:
          $ref: "#/components/schemas/ModelSpec"
        notes:
          type: string
        hasReferenceFace:
          description: 是否已上传参考脸；有则该脸是后续生成的唯一身份基准。
          type: boolean
        referenceFaceUrl:
          type: string
        selectedPortrait:
          anyOf:
            - $ref: "#/components/schemas/ModelPortrait"
            - type: "null"
        portraitCount:
          minimum: 0
          type: integer
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
    ModelList:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EcomModel"
        nextCursor:
          anyOf:
            - type: string
            - type: "null"
    ModelPortraitList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelPortrait"
    CreateModelInput:
      type: object
      required:
        - name
        - spec
      properties:
        name:
          minLength: 1
          maxLength: 40
          type: string
        spec:
          $ref: "#/components/schemas/ModelSpec"
        notes:
          maxLength: 2000
          type: string
    UpdateModelInput:
      type: object
      properties:
        name:
          minLength: 1
          maxLength: 40
          type: string
        spec:
          $ref: "#/components/schemas/ModelSpec"
        notes:
          maxLength: 2000
          type: string
    CreateModelCastJobInput:
      description: 发起一次选角生成：按模特当前 spec 编译定妆照 prompt，产出 candidateCount 张候选。
      type: object
      required:
        - providerId
        - imageModelId
        - aspectRatio
      properties:
        providerId:
          format: uuid
          type: string
        imageModelId:
          minLength: 1
          type: string
        aspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        imageResolution:
          description: 输出分辨率档位；仅当所选模型支持（Seedream/Gemini 3）时生效，gpt-image 固定 1K 输出。
          $ref: "#/components/schemas/ImageResolution"
        quality:
          description: 生图质量档位；仅 gpt-image 家族消费，其余模型忽略。缺省 high。
          $ref: "#/components/schemas/ImageQuality"
        outputFormat:
          description: 产物图片格式；仅支持的模型（gpt-image / Seedream 4.5+）会收到该参数，缺省 png。
          $ref: "#/components/schemas/ImageOutputFormat"
        candidateCount:
          minimum: 1
          maximum: 4
          default: 1
          type: integer
...